refactor(lib): organize files into `client` and `server` folders
