+++ index.d.ts
@@ -1,3 +1,3 @@
 // TypeScript Version: 4.1
 
-export { default } from './lib/html-to-dom-server';
+export { default } from './lib/server/html-to-dom';
