+++ lib/client/constants.d.ts
@@ -3,6 +3,6 @@
 /**
  * SVG elements, unlike HTML elements, are case-sensitive.
  *
- * @see {@link https://developer.mozilla.org/docs/Web/SVG/Element#SVG_elements_A_to_Z}
+ * {@link https://developer.mozilla.org/docs/Web/SVG/Element#SVG_elements_A_to_Z}
  */
 export const CASE_SENSITIVE_TAG_NAMES: string[];
