+++ lib/client/utilities.js
