+++ test/server/server.js
@@ -2,7 +2,7 @@ const { assert } = require('chai');
 const htmlparser = require('htmlparser2');
 const cases = require('../cases');
 const { runTests, throwErrors } = require('../helpers');
-const { CASE_SENSITIVE_TAG_NAMES } = require('../../lib/constants');
+const { CASE_SENSITIVE_TAG_NAMES } = require('../../lib/client/constants');
 
 /**
  * Tests case-sensitive tags (SVG) to make sure their case is preserved.
@@ -32,7 +32,7 @@ describe('client parser in jsdom', () => {
   // before
   const jsdomify = require('jsdomify').default;
   jsdomify.create();
-  const clientParser = require('../../lib/html-to-dom-client');
+  const clientParser = require('../../lib/client/html-to-dom');
 
   // tests
   throwErrors(assert, clientParser);
